+++ delta-coverage-core/src/main/kotlin/io/github/surpsg/deltacoverage/config/DeltaCoverageConfig.kt
@@ -138,24 +138,21 @@ enum class CoverageEntity {
 class ReportsConfig private constructor(
     val html: ReportConfig,
     val xml: ReportConfig,
-    val csv: ReportConfig,
     val baseReportDir: String,
     val fullCoverageReport: Boolean,
 ) {
 
-    override fun toString(): String = "ReportsConfig(html=$html, xml=$xml, csv=$csv" +
+    override fun toString(): String = "ReportsConfig(html=$html, xml=$xml" +
             ", baseReportDir='$baseReportDir', fullCoverageReport=$fullCoverageReport)"
 
     @DeltaCoverageConfigMarker
     class Builder internal constructor() {
         var html: ReportConfig = ReportConfig {}
         var xml: ReportConfig = ReportConfig {}
-        @Deprecated(message = "This property will be removed in the next major release.")
-        var csv: ReportConfig = ReportConfig {}
         var baseReportDir: String = ""
         var fullCoverageReport: Boolean = false
 
-        fun build(): ReportsConfig = ReportsConfig(html, xml, csv, baseReportDir, fullCoverageReport)
+        fun build(): ReportsConfig = ReportsConfig(html, xml, baseReportDir, fullCoverageReport)
     }
 
     companion object {