+++ delta-coverage-core/src/main/kotlin/io/github/surpsg/deltacoverage/report/Reports.kt
@@ -35,7 +35,6 @@ data class JacocoReport(
 enum class ReportType {
     HTML,
     XML,
-    @Deprecated("CSV will be removed soon.")
     CSV,
 }